Discord bot application Basic tasks: Bot Working
Can read messages and can reply to it.

Task
Read a message in one of the Channel(# reminders) with role
Decide which category does the member belong
Make a list
Send it to anyone with role ( remind).

Solution:

In a channel-OCEM people who attended a meeting

Should read a role of members.
Categorize them and make a list of it.

Executed until now